прогрессбары из релиза

+++ xr_3da/xrGame/ui/UIMotionIcon.cpp
@@ -75,18 +75,18 @@ void CUIMotionIcon::ShowState(EState state)
 	m_curren_state=state;
 }
 
-void CUIMotionIcon::SetPower(s16 Pos)
+void CUIMotionIcon::SetPower(float Pos)
 {
 	m_power_progress.SetProgressPos(Pos);
 }
 
-void CUIMotionIcon::SetNoise(s16 Pos)
+void CUIMotionIcon::SetNoise(float Pos)
 {
 	Pos	= clampr(Pos, m_noise_progress.GetRange_min(), m_noise_progress.GetRange_max());
 	m_noise_progress.SetProgressPos(Pos);
 }
 
-void CUIMotionIcon::SetLuminosity(s16 Pos)
+void CUIMotionIcon::SetLuminosity(float Pos)
 {
 	Pos	= clampr(Pos, m_luminosity_progress.GetRange_min(), m_luminosity_progress.GetRange_max());
 	m_luminosity_progress.SetProgressPos(Pos);
@@ -106,7 +106,7 @@ void CUIMotionIcon::Update()
 	inherited::Update();
 }
 
-void CUIMotionIcon::SetActorVisibility		(u16 who_id, s16 value)
+void CUIMotionIcon::SetActorVisibility		(u16 who_id, float value)
 {
 	value	= clampr(value, m_luminosity_progress.GetRange_min(), m_luminosity_progress.GetRange_max());
